EDM-885: RPM versioning

+++ hack/build_rpms.sh
@@ -8,10 +8,13 @@ if [ -n "${FLIGHTCTL_RPM:-}" ]; then
 fi
 
 # our RPM build process works in rpm bases systems so we wrap it if necessary
-if ! command -v packit >/dev/null 2>&1 ]; then
+if ! command -v packit >/dev/null 2>&1; then
     echo "Building RPMs on a system without packit, using container"
     cat >bin/build_rpms.sh <<EOF
-dnf install -y go-rpm-macros || exit 1
+if ! dnf install -y go-rpm-macros; then
+    echo "Failed to install go-rpm-macros package"
+    exit 1
+fi
 ./hack/build_rpms_packit.sh
 EOF
     podman run --privileged --rm -t -v "$(pwd)":/work quay.io/flightctl/ci-rpm-builder:latest bash /work/bin/build_rpms.sh
